Add 'parameters' argument to ISAMAppliance.get()
The requests library supports 'params', a dictionary of query string
parameters.

Eg passing params={"foo": "bar baz"} in a GET request to /the/endpoint
will query /the/endpoint?foo=bar%20baz.

I've added it to get() because it's required for authentication policy
queries 'sortBy' and 'filter', and some others in the RAPI, including
authentication mechanisms.