+1 feature: set EmailMessage, ScimConfig, CI server connections properties by name
This feature attempts to find the uuids from the properties "EmailMessage.serverConnection", "ScimConfig.serverConnection" and "CI.serverConnection". If a UUID is found, the value for this property is replaced by this UUID, otherwise the value remains unchanged.
This allows you to set ServerConnections by name, while still allowing you to create the ServerConnection directly from the UUID.