Fix: unnecessary parameter id on import_file id now optional
appliance REST api does not support the `id` parameter on the import of transformation rules. Therefore it is not necessary to insist on that parameter on the import_file function.
For backwards compatibility the id parameter is now defined with a default value of None, but as it is not used afterwards the id parameter is removed from the _check_import function.